rename: components(pages)폴더 추가
