+++ src/app/(pages)/signup/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import SignupPage from "@/app/_components/template/SignupPage";
+import SignupPage from "@/app/_components/pages/SignupPage";
 import React from "react";
 
 export default function page() {
