+++ src/app/_components/pages/HomePage.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import HomeTemplate from "../template/HomeTemplate";
+
+export default function HomePage() {
+  return <HomeTemplate />;
+}
