+++ src/app/_components/pages/LoginPage.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import LoginTemplate from "../template/LoginTemplate";
+
+export default function LoginPage() {
+  return <LoginTemplate />;
+}
