+++ src/app/_components/pages/SignupPage.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import SignupTemplate from "../template/SignupTemplate";
+
+export default function SignupPage() {
+  return <SignupTemplate />;
+}
