--- src/app/_components/template/LoginPage.tsx
@@ -1,11 +1,0 @@
-import { PageWrapper } from "@/app/_styles/common.style";
-import React from "react";
-import LoginCard from "../organism/group/LoginCard";
-
-export default function LoginPage() {
-  return (
-    <PageWrapper>
-      <LoginCard />
-    </PageWrapper>
-  );
-}
