--- src/app/_components/template/SignupPage.tsx
@@ -1,11 +1,0 @@
-import { PageWrapper } from "@/app/_styles/common.style";
-import React from "react";
-import SignupCard from "../organism/group/SingupCard";
-
-export default function SignupPage() {
-  return (
-    <PageWrapper>
-      <SignupCard />
-    </PageWrapper>
-  );
-}
